requirement.txt
asgiref==3.8.1
captcha==0.6.0
Django==5.1.6
django-cors-headers==4.7.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.1